requirements.txt
streamlit
streamlit-option-menu
streamlit-lottie
google-generativeai>=0.5.0
requests
joblib
numpy
pandas
pillow
scikit-learn
xgboost
lightgbm
catboost
onnxruntime
google-cloud-aiplatform
vertexai
